% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/display_pbi_palette.R
\name{display_pbi_pal}
\alias{display_pbi_pal}
\title{Display Microsoft Power BI Palettes}
\usage{
display_pbi_pal(palette = "classic",
  n = length(pbi_palettes[[palette]]))
}
\arguments{
\item{palette}{A palette name from \code{names(pbi_palettes)}, defaults to
\code{default} palette}

\item{n}{Number of colours from the palette to display}
}
\value{
Plot of colours
}
\description{
Display Microsoft Power BI Palettes
}
\examples{
display_pbi_pal("classic", 5)

}
\author{
Keyur Shah
}
