% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pbi_pal.R
\name{pbi_pal}
\alias{pbi_pal}
\title{Microsoft PowerBI palettes}
\usage{
pbi_pal(palette = "default", direction = 1)
}
\arguments{
\item{palette}{A palette name from \code{names(pbi_palettes)}, defaults to
\code{default} palette}

\item{direction}{Sets the order of colours in the scale. If -1, the order of
colours is reversed. Defaults to 1}
}
\value{
A function with paramater \code{n}, number of colors required from
 the palette \code{name} palette
}
\description{
Creates nice looking color palettes from Microsoft PowerBI

\code{pbi_pal} makes the color palettes from Microsoft PowerBI available
 as R palettes. Names of 8 palettes are as follows:
 \enumerate{
  \item \code{default}
  \item \code{city park}
  \item \code{classroom}
  \item \code{colorblind safe}
  \item \code{electric}
  \item \code{high contrast}
  \item \code{sunset}
  \item \code{twilight}
 }
}
\examples{
pbi_pal("default", direction = 1)(10)

}
\author{
Keyur Shah
}
