% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{pbitheme}
\alias{pbitheme}
\alias{theme_pbi}
\title{Complete themes}
\usage{
theme_pbi(base_size = 11, base_family = "",
  base_line_size = base_size/22, base_rect_size = base_size/22)
}
\arguments{
\item{base_size}{base font size}

\item{base_family}{base font family}

\item{base_line_size}{base size for line elements}

\item{base_rect_size}{base size for rect elements}
}
\description{
These are complete themes from Microsoft Power BI which control
all non-data display.
}
\details{
\describe{
   \item{theme_pbi}{A theme similar to one shown in Microsoft Power BI}
}
}
\examples{
if (require(ggplot2)) {
   mtcars2 <- within(mtcars, {
    vs <- factor(vs, labels = c("V-shaped", "Straight"))
    am <- factor(am, labels = c("Automatic", "Manual"))
    cyl <- factor(cyl)
    gear <- factor(gear)
   })

  p1 <- ggplot(mtcars2) +
    geom_point(aes(x = wt, y = mpg, colour = gear)) +
    labs(title = "Fuel economy declines as weight increases",
         subtitle = "(1973-74)",
         caption = "Data from the 1974 Motor Trend US magazine.",
         tag = "Figure 1",
         x = "Weight (1000 lbs)",
         y = "Fuel economy (mpg)",
         colour = "Gears")

  p1 + theme_pbi()
}

# Theme examples with panels

if (require(ggplot2)) {
  p2 <- p1 + facet_grid(vs ~ am)

  p2 + theme_pbi()
}

}
