% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{scale_color_pbi}
\alias{scale_color_pbi}
\alias{scale_colour_pbi}
\alias{scale_fill_pbi}
\title{Color and fill scales from Microsoft PowerBI}
\usage{
scale_color_pbi(..., palette = "classic", direction = 1,
  aesthetics = "colour")

scale_colour_pbi(..., palette = "classic", direction = 1,
  aesthetics = "colour")

scale_fill_pbi(..., palette = "classic", direction = 1,
  aesthetics = "fill")
}
\arguments{
\item{...}{Other arguments passed on to \code{\link[ggplot2]{discrete_scale}}}

\item{palette}{A palette name from \code{names(pbi_palettes)}}

\item{direction}{Sets the order of colours in the scale. If -1, the order of colours
is reversed. Defaults to 1}

\item{aesthetics}{Character string or vector of character strings listing
the name(s) of the aesthetic(s) that this scale works with. This can be
useful, for example, to apply colour settings to the \code{colour} and
\code{fill} aesthetics at the same time via
\code{aesthetics = c("colour", "fill")}}
}
\description{
The \code{pbi} scales provide default, city park, classroom,
colorblind safe, electric, high contrast, sunset and twilight color schemes
from Microsoft Power BI. These are well suited to display discrete values.

The Microsoft PowerBI scales are not designed to be extended to continuous
data.
}
\examples{

if (require(ggplot2)) {
  ggplot(iris, aes(x = Sepal.Length, y = Sepal.Width, color = Species)) +
   geom_point(size = 4) +
   theme_bw(base_family = "mono") +
   scale_color_pbi(palette = "twilight")

  ggplot(mpg, aes(x = manufacturer, fill = manufacturer)) +
   geom_bar() +
   scale_fill_pbi(palette = "colorblind safe") +
   theme_bw(base_family = "serif") +
   labs(x = "Manufacturer",
        y = "Count",
        title = "Manufacturer Counts",
        subtitle = "Dodge is highest!")
}

}
